requirements.txt
streamlit>=1.24.0
pandas>=2.0.0
plotly>=5.14.0
beautifulsoup4==4.12.2
requests>=2.31.0
SQLAlchemy>=2.0.0
python-dotenv>=1.0.0
numpy==1.26.4
APScheduler==3.10.4
google-api-python-client>=2.88.0
